[fix]: 修复BUG;节点数上的每一个mesh的ID会因为点击展开之后发生改变 --修正为 每个mesh都有各自独有的ID

+++ HW3D/App.cpp
@@ -106,7 +106,7 @@ void App::DoFrame()
 	{	// 且光标禁用(F1状态)
 		if (!wnd.CursorEnabled())
 		{	
-			cam.Rotate(delta->x, delta->y);// 对摄像机执行按鼠标输入分量的旋转
+			cam.Rotate( (float)delta->x, (float)delta->y);// 对摄像机执行按鼠标输入分量的旋转
 		}
 	}
